simplifies code
